// src/app/Models/model.ts
export class Perfil {

    avatarUrl: string='';
    birthDate: Date = new Date();
    country: string ='';
    email: string ='';
    firstName: string='';
    lastName: string ='';
    phoneNumber: string='';
    state: string='';
    userName: string='';

}
export class token{
    atoken: string = '';
}

export class Pefil1 {
   nombre:String='';
   apellido:String='';
   ciudad:String='';
   email:String='';
   fechaNac:Date = new Date();
   id:number=0;
   pais:String='';
   password:String='';
   numTel:number=0;
}

// apellido
// : 
// "ortega"
// ciudad
// : 
// "cord"
// email
// : 
// "tomi"
// fechaNac
// : 
// 1703100321161
// id
// : 
// 1
// nombre
// : 
// "tomas"
// numTel
// : 
// 2
// pais
// : 
// "arg"
// password
// : 
// "tomi"